LPS-178954 Moves method to display context, and calculate it only when it is necessary
