+++ modules/apps/segments/segments-web/src/main/java/com/liferay/segments/web/internal/portlet/SegmentsPortlet.java
@@ -72,28 +72,13 @@ public class SegmentsPortlet extends MVCPortlet {
 		throws IOException, PortletException {
 
 		renderRequest.setAttribute(
-			SegmentsWebKeys.EXCLUDED_ROLE_NAMES, _getExcludedRoleNames());
-
-		renderRequest.setAttribute(
 			SegmentsDisplayContext.class.getName(),
 			new SegmentsDisplayContext(
 				_analyticsSettingsManager, _groupLocalService, _itemSelector,
 				_language, _portal, renderRequest, renderResponse,
-				_segmentsConfigurationProvider, _segmentsEntryService));
+				_roleTypeContributorProvider, _segmentsConfigurationProvider, _segmentsEntryService));
 
 		super.render(renderRequest, renderResponse);
-	}
-
-	private String[] _getExcludedRoleNames() {
-		RoleTypeContributor roleTypeContributor =
-			_roleTypeContributorProvider.getRoleTypeContributor(
-				RoleConstants.TYPE_SITE);
-
-		if (roleTypeContributor != null) {
-			return roleTypeContributor.getExcludedRoleNames();
-		}
-
-		return new String[0];
 	}
 
 	@Reference
